requirements.txt
folium==0.17.0
GeoAlchemy2==0.15.2
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
greenlet==3.0.3
mapclassify==2.7.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
pandas==2.2.1
pillow==10.4.0
psycopg-binary==3.2.1
psycopg2==2.9.9
python-dotenv==1.0.1
shapely==2.0.5
SQLAlchemy==2.0.31
tzdata==2024.1
